requirements.txt
et-xmlfile==1.1.0
numpy==1.24.2
openpyxl==3.1.2
pandas==1.5.3
Pillow==9.4.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
six==1.16.0
